Refactor: #110 Move CreditCardComponent to SignUpPage Directory
